# summary simplification, test cases, and eager checks