--- a/.github/workflows/permanent.yml
+++ b/.github/workflows/permanent.yml
@@ -3,14 +3,13 @@ on:
   push:
     branches:
       - 'master'
+  pull_request:
 jobs:
   document:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@latest
-        with:
-          version: '1.3'
       - uses: julia-actions/julia-docdeploy@releases/v1
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}