--- a/docs/make.jl
+++ b/docs/make.jl
@@ -1,5 +1,5 @@
 using JuliennedArrays
 using Documenter: deploydocs, makedocs
 
-makedocs(sitename = "JuliennedArrays.jl", modules = [JuliennedArrays], doctest = false)
+makedocs(sitename = "JuliennedArrays.jl", modules = [JuliennedArrays], doctest = true, strict=true)
 deploydocs(repo = "github.com/bramtayl/JuliennedArrays.jl.git")