--- a/src/JuliennedArrays.jl
+++ b/src/JuliennedArrays.jl
@@ -3,6 +3,9 @@ module JuliennedArrays
 import Base: axes, getindex, setindex!, size
 using Base: @pure, tail
 
+export Slices, Align
+export True, False
+
 @inline is_in(needle::Needle, straw1::Needle, straws...) where {Needle} = True()
 @inline is_in(needle, straw1, straws...) = is_in(needle, straws...)
 @inline is_in(needle) = False()
@@ -33,9 +36,6 @@ struct False <: TypedBool end
 @inline not(::False) = True()
 @inline not(::True) = False()
 
-export True
-export False
-
 @inline getindex_unrolled(into::Tuple{}, switches::Tuple{}) = ()
 @inline function getindex_unrolled(into, switches)
     next = getindex_unrolled(tail(into), tail(switches))
@@ -54,9 +54,16 @@ end
         first(old), setindex_unrolled(tail(old), new, tail(switches))...
     end
 
+###
+# Slices
+###
 struct Slices{Item,Dimensions,Whole,Alongs} <: AbstractArray{Item,Dimensions}
     whole::Whole
     alongs::Alongs
+    function Slices{T,N,W,A}(whole::W, alongs::A) where {T,N,W,A}
+        any(isequal(True()), alongs) || throw(DimensionMismatch("Expected to have at least one active slicing dimension."))
+        new{T,N,W,A}(whole, alongs)
+    end
 end
 @inline Slices{Item,Dimensions}(
     whole::Whole,
@@ -74,12 +81,8 @@ end
 @inline setindex!(slices::Slices, value, indices::Int...) =
     slices.whole[slice_index(slices, indices)...] = value
 
-@inline axis_or_1(switch, axis) =
-    if untyped(switch)
-        axis
-    else
-        1
-    end
+@inline axis_or_1(switch, axis) = untyped(switch) ? axis : 1
+
 """
     Slices(whole, alongs::TypedBool...)
 
@@ -115,13 +118,14 @@ julia> size(first(larger_slices))
 (5,)
 ```
 """
-Slices(whole::AbstractArray, alongs::TypedBool...) = Slices{
-    typeof(@inbounds view(whole, map(axis_or_1, alongs, axes(whole))...)),
-    length(getindex_unrolled(alongs, map(not, alongs))),
-}(
-    whole,
-    alongs,
-)
+function Slices(whole::AbstractArray, alongs::TypedBool...)
+    length(alongs) <= ndims(whole) || throw(ArgumentError("$(length(alongs)) dimensions are specified, expected to be <= $(ndims(whole))"))
+    # mark all tailing dimensions as outer dimensions
+    alongs = (alongs..., ntuple(i->False(), ndims(whole)-length(alongs))...)
+    x = @inbounds view(whole, map(axis_or_1, alongs, axes(whole))...)
+    N = length(getindex_unrolled(alongs, map(not, alongs)))
+    return Slices{typeof(x),N}(whole, alongs)
+end
 
 """
     Slices(whole, alongs::Int...)
@@ -152,16 +156,25 @@ julia> map(sum, s)
  22
 ```
 """
-Slices(
-    whole::AbstractArray{Item,NumberOfDimensions},
-    alongs::Int...,
-) where {Item,NumberOfDimensions} =
-    Slices(whole, in_unrolled(as_vals(alongs...), ntuple(Val, NumberOfDimensions)...)...)
-export Slices
+function Slices(whole::AbstractArray{T,N}, alongs::Int...) where {T,N}
+    any(x->x>N, alongs) && throw(ArgumentError("All alongs values $(alongs) should be less than $(N)"))
+    Slices(whole, in_unrolled(as_vals(alongs...), ntuple(Val, N)...)...)
+end
 
+###
+# Align
+###
 struct Align{Item,Dimensions,Sliced,Alongs} <: AbstractArray{Item,Dimensions}
     slices::Sliced
     alongs::Alongs
+    function Align{T,N,S,A}(slices::S, alongs::A) where {T,N,S,A}
+        sz = size(first(slices))
+        all(x->sz==size(x), slices) || throw(ArgumentError("All sizes of slices should be the same."))
+        length(alongs) == N || throw(DimensionMismatch("The total dimension $(N) is expected to be the sum of inner dimension $(length(sz)) and outer dimension $(length(alongs))"))
+        inner_dimensions = mapreduce(isequal(True()), +, alongs)
+        inner_dimensions == ndims(first(slices)) || throw(DimensionMismatch("Only $inner_dimensions inner dimensions are used, expected $(ndims(first(slices))) dimensions."))
+        new{T,N,S,A}(slices, alongs)
+    end
 end
 @inline Align{Item,Dimensions}(
     slices::Sliced,
@@ -221,7 +234,6 @@ julia> slices
     alongs::TypedBool...,
 ) where {Item,InnerDimensions,OuterDimensions} =
     Align{Item,OuterDimensions + InnerDimensions}(slices, alongs)
-export Align
 
 """
     Along(slices, alongs::Int...)