Because the current test fails on the master version, I submit all commits in one single PR.

Changes:

- simplify summary show of `Align` and `Slices`
- add more eager argument checks on `Align` and `Slices` (closes #25 )
- add comprehensive testsets
- some style changes

A new feature:

- allow incomplete `Slices` construction: `Slices(Xs, True())` is equivalent to `Slices(Xs, True(), False(), ... False())`. This feature will be ambiguous for `Align` so that's not allowed.