@bramtayl 2021-08-25
Hi! Thanks for the contribution! I think most of the changes look good. Could you separate out the changes to the constructors to a separate pull request? That's the part I'm hesitant about