@bramtayl 2021-08-25
> allow incomplete Slices construction: Slices(Xs, True()) is equivalent to Slices(Xs, True(), False(), ... False())

I'm not sure about this. `Slices(Xs, True())` seems unclear for many-dimensional arrays