@johnnychen94 2021-08-25
There are several other things that can be improved, but for simplicity, I'd like to leave them as future PRs:

- For type inferrable and compile-time optimization. Maybe introduce Static.jl, it has a more sophisticated design and also supports common number operations. The drawback is the loading time is 10x larger than this package.
- There can be some optimizations on `Align(Slices(...)...)` and `Slices(Align(...)...)`.

---

For runtime overhead, I wouldn't say it's zero overhead, but I think they're necessary checks.

```julia
using BenchmarkTools
using JuliennedArrays

# Align

Xs = [rand(3, 4) for _ in 1:8];
Xs = reshape(Xs, 2, 4);

@btime Align($Xs, 1, 3);
# PR: 998.615 ns (2 allocations: 48 bytes)
# master: 1.005 μs (2 allocations: 48 bytes)
@btime Align($Xs, $(True()), $(False()), $(True()), $(False()));
# PR: 1.475 ns (0 allocations: 0 bytes)
# master: 1.431 ns (0 allocations: 0 bytes)

X = rand(2, 3, 4, 5);
Xs = Slices(X, True(), False(), True(), False());

@btime Align($Xs, 1, 3);
# PR: 1.008 μs (3 allocations: 64 bytes)
# master: 995.364 ns (3 allocations: 64 bytes)
@btime Align($Xs, $(True()), $(False()), $(True()), $(False()));
# PR: 6.798 ns (0 allocations: 0 bytes)
# master: 1.432 ns (0 allocations: 0 bytes)

# Slices

X = rand(2, 3, 4, 5);
@btime Slices($X, 1, 3);
# PR: 2.731 μs (19 allocations: 816 bytes)
# master: 2.745 μs (18 allocations: 784 bytes)
@btime Slices($X, $(True()), $(False()), $(True()), $(False()));
# PR: 1.620 ns (0 allocations: 0 bytes)
# master: 1.432 ns (0 allocations: 0 bytes)

Xs = [rand(3, 4) for _ in 1:8];
Xs = reshape(Xs, 2, 4);
X = Align(Xs, True(), False(), True(), False());
@btime Slices($X, 1, 3);
# PR: 2.923 μs (20 allocations: 832 bytes)
# master: 2.871 μs (19 allocations: 800 bytes)
@btime Slices($X, $(True()), $(False()), $(True()), $(False()));
# PR: 1.831 μs (18 allocations: 768 bytes)
# master: 1.742 μs (18 allocations: 768 bytes)

```